Enqueue Workers::ProcessPagerDutyEvent with 5 second delay
Sometimes slackerduty will fetch a PagerDuty incident, before PagerDuty
has associated the alert to it, this results in not being able to show
richer information related to an incident.

This isn't an ideal solution, but it is easy. We could alternatively
enqueue something to update an incident after a couple of seconds.